requirements.txt
Flask-SQLAlchemy==2.0
MySQL-python==1.2.5
request==0.0.2
requests==2.8.1
SQLAlchemy==1.0.8
zope.interface==4.1.3
